refactor: 1. change Newsletter.tsx to server component 2. set sideNavbar Link's attribute preFetch to true
